Changes in api.py and action_service.py
Modified Api.py to segregate Constants and logger methods. This will help in keeping the functions modular and if we have to add another resource to the existing two, they can reuse the same methods. Removed redundancy in code for status checking. Used the Util as an parent to the resource classes so that assignments can be refered via SUPER.
In action_service; assuming actions is a list/json, changed the statement to check for empty and Null object before proceeding further. Previously it would have run for a non null but empty list.